+++ tests/utils.ts
@@ -3,3 +3,40 @@ import BigNumber from "bignumber.js";
 export const convertToUnit = (amount: string | number, decimals: number) => {
   return new BigNumber(amount).times(new BigNumber(10).pow(decimals)).toString();
 };
+
+const initMainnetUser = async (user: string) => {
+  await impersonateAccount(user);
+  return ethers.getSigner(user);
+};
+
+export async function setForkBlock(blockNumber: number) {
+  await network.provider.request({
+    method: "hardhat_reset",
+    params: [
+      {
+        forking: {
+          jsonRpcUrl: process.env.BSC_ARCHIVE_NODE,
+          blockNumber: blockNumber,
+        },
+      },
+    ],
+  });
+}
+
+const forking = (blockNumber: number, fn: () => void) => {
+  describe(`riskFundConverter #${blockNumber}`, () => {
+    before(async () => {
+      await setForkBlock(blockNumber);
+    });
+    fn();
+  });
+};
+
+async function impersonateAccount(accountAddress) {
+  await hre.network.provider.request({
+    method: "hardhat_impersonateAccount",
+    params: [accountAddress],
+  });
+}
+
+export { initMainnetUser, forking, impersonateAccount };
